// B2CReplacementDesigner.Client/src/types/trace.ts
export interface JourneyContext {
    /** The journey ID (UserJourney or SubJourney) */
    journeyId: string;
    /** Human-readable name of the journey */
    journeyName: string;
    /** The last seen ORCH_CS in this context */
    lastOrchStep: number;
    /** The entry timestamp */
    entryTimestamp: Date;
}